requirements.txt
#click==8.0.3
Flask==2.0.2
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
#importlib-metadata==4.8.2
#itsdangerous==2.0.1
#Jinja2==3.0.3
#MarkupSafe==2.0.1
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
SQLAlchemy==1.4.27
Flask-Migrate==3.1.0
psycopg2-binary
sodapy==2.1.0
PyYAML==5.4.1
flask-marshmallow